Renaming: 'AreaSelection' <-> 'LayerSelection'
