+++ frontend/src/src/components/AreaSelection.svelte
@@ -1,15 +1,14 @@
 <script>
   import {onMount} from 'svelte';
+  import '../leaflet_components/L.TileLayer.NutsLayer.js';
+  import '../leaflet_components/L.DrawingLayer.js';
   import '../leaflet_components/L.TileLayer.QueryableLayer.js';
-  import {getGeofiles, getLegend, getLayerType, getOpenairLink, WMS_URL} from '../client.js';
-  import {activeOverlayLayersStore} from '../stores.js';
+  import {getGeofiles, WMS_URL} from '../client.js';
+  import {activeSelectionLayerStore} from '../stores.js';
 
-  const activeOverlayLayers = [];
-  let overlayLayers = [];
-  let isLayerListReady = false;
-  let overlayLayersFilter = '';
-  let filteredOverlayLayers = [];
-
+  // List of queryable layers that are used as selection layers.
+  // The order in which they appear is mirrored in the order the layers
+  // are displayed.
   export const SELECTIONS_LIST= [
     'country.geojson',
     'NUTS1.geojson',
@@ -19,29 +18,15 @@
   ];
   export const SELECTIONS = new Set(SELECTIONS_LIST);
 
+  let selectionLayers = [];
+  let isLayerListReady = false;
 
-  function splitName(name) {
-    return name.substring(3).replace(/\.[^/.]+$/, '');
-  };
-
-  function toQueryableLayer(layerName) {
-    const layer = L.tileLayer.queryableLayer(
+  function toNutsLayer(layerName) {
+    const layer = L.tileLayer.nutsLayer(
         WMS_URL,
         {
           transparent: 'true',
-          layers: encodeURIComponent(layerName),
-          format: 'image/png',
-        },
-    );
-    return layer;
-  }
-
-  function toOverlayLayer(layerName) {
-    const layer = L.tileLayer.wms(
-        WMS_URL,
-        {
-          transparent: 'true',
-          layers: encodeURIComponent(layerName),
+          layers: layerName,
           format: 'image/png',
         },
     );
@@ -53,64 +38,57 @@
     for (const [layer, layerParameters] of Object.entries(layers)) {
       let leafletLayer;
       console.log(layer, layerParameters);
-      if (!SELECTIONS.has(layer)) {
-        const legend = getLegend(layer);
-        console.log(legend);
-        const layerType = getLayerType(layer);
-        console.log(layerType);
-
-        const openairLink = getOpenairLink(layer);
-        console.log(openairLink);
-
-        if (layerParameters.isQueryable) {
-          leafletLayer = toQueryableLayer(layer);
-        } else {
-          leafletLayer = toOverlayLayer(layer);
-        }
-
-        leafletLayer.name = layer;
-        leafletLayer.datasetId = parseInt(layer.substring(0, 2));
-        leafletLayer.legend_promise = legend;
-        leafletLayer.openairLink_promise = openairLink;
-        leafletLayer.layer_type_promise = layerType;
-        overlayLayers.push(leafletLayer);
-
-        // select if dataset to show is set in the URL
-        const urlParams = new URLSearchParams(window.location.search);
-        if (urlParams.has('datasetId')) {
-          const datasetSelected = parseInt(urlParams.get('datasetId'));
-          if (leafletLayer.datasetId == datasetSelected) {
-            leafletLayer.checked = true;
-            activeOverlayLayers.push(leafletLayer);
+      if (SELECTIONS.has(layer)) {
+        // We can put something else than the full name of the file
+        function convertName(layer) {
+          if (layer == 'country.geojson') {
+            return 'Country';
+          } else if (layer == 'NUTS1.geojson') {
+            return 'Region NUTS1';
+          } else if (layer == 'NUTS2.geojson') {
+            return 'Region NUTS2';
+          } else if (layer == 'NUTS3.geojson') {
+            return 'Region NUTS3';
+          } else if (layer == 'LAU.geojson') {
+            return 'Cities';
+          } else {
+            return layer;
           }
-        }
+        };
+        leafletLayer = toNutsLayer(layer);
+        leafletLayer.name = convertName(layer);
+        // selection go on top
+        leafletLayer.setZIndex(1000);
+        selectionLayers.push(leafletLayer);
       }
     }
 
-    // we can now push the potential selected dataset
-    $activeOverlayLayersStore = activeOverlayLayers;
-
-    overlayLayers.sort(function(layer0, layer1) {
-      const name0 = splitName(layer0.name);
-      const name1 = splitName(layer1.name);
-      if (name0 < name1) {
-        return -1;
-      }
-      if (name0 > name1) {
-        return 1;
-      }
-      return 0;
+    selectionLayers.sort( function(layer0, layer1) {
+      const areaList = [
+        'Country',
+        'Region NUTS1',
+        'Region NUTS2',
+        'Region NUTS3',
+        'Cities',
+      ];
+      const a = areaList.indexOf(layer0.name);
+      const b = areaList.indexOf(layer1.name);
+      return a - b;
     });
 
-    overlayLayers = overlayLayers;
-    filteredOverlayLayers = overlayLayers;
+    const drawingLayer = getDrawingLayer();
+    drawingLayer.name = 'Selection';
+    selectionLayers.push(drawingLayer);
+    selectionLayers = selectionLayers;
     isLayerListReady = true;
   });
 
+  function getDrawingLayer() {
+    return new L.DrawingLayer();
+  }
+
   $: {
-    console.log('layer changed in selector to ' + $activeOverlayLayersStore);
-    filteredOverlayLayers = overlayLayers.filter((layer) =>
-      layer.name.toLowerCase().indexOf(overlayLayersFilter.toLowerCase()) !== -1);
+    console.log('layer changed in selector to ' + $activeSelectionLayerStore);
   }
 </script>
 
@@ -121,7 +99,7 @@
     padding: 4px;
     border: 1px solid #27275b;
     border-radius: 0px;
-    background-color: #eff4fa;
+    background-color: #eff4fa !important;
     box-sizing: border-box;
   }
 
@@ -140,13 +118,9 @@
     border : none;
   }
 
-  #overlay_layers {
-    max-height: 300px;
+  #selection_layers {
     overflow-y: auto;
     border : none;
-    overflow-y: scroll;
-    scrollbar-color: #27275b;
-    scrollbar-width: thin;
   }
 
   label {
@@ -157,84 +131,20 @@
     overflow-x: hidden;
     margin-top: 2px;
   }
-  .overlay_search {
-    width: 100%;
-  }
-
-  .box {
-    height: 10px;
-    width: 10px;
-    border: 1px solid black;
-    display: inline-block;
-  }
-
-  #metadata_box {
-    border: 1px solid #27275b;
-    border-radius: 0px;
-    background-color: #fff;
-    padding: 5px;
-    width: inherit;
-    margin-right: 10px;
-    margin-left: 10px;
-    overflow: hidden;
-  }
 </style>
-
 
 <div id="map_selection" on:click|stopPropagation on:dblclick|stopPropagation on:wheel|stopPropagation>
   {#if !isLayerListReady}
     Loading layers...
   {:else}
-    <h3>Overlays layers</h3>
-    <input bind:value={overlayLayersFilter} class="overlay_search" placeholder="Search layer...">
-
-    <div id="overlay_layers" style="margin-top: 10px;">
-      {#each filteredOverlayLayers as overlayLayer (overlayLayer.name)}
-        <label title={splitName(overlayLayer.name)}>
-          <input type=checkbox bind:group={$activeOverlayLayersStore} value={overlayLayer} bind:checked={overlayLayer.checked}>
-            {splitName(overlayLayer.name)}
+    <h3>Area selection</h3>
+    <div id="selection_layers">
+      {#each selectionLayers as selectionLayer}
+        <label title={selectionLayer.name}>
+          <input type=radio bind:group={$activeSelectionLayerStore} value={selectionLayer}>
+          {selectionLayer.name}
         </label>
-
-        <div id="metadata_box" hidden={!overlayLayer.checked}>
-          {#await overlayLayer.legend_promise}
-            <div>...waiting for legend</div>
-          {:then legend}
-            <div><b>{legend.variable.variable}</b></div>
-            <div>
-              {#each legend.style as color}
-                <div style="display: inline-block;">
-                  {#await overlayLayer.layer_type_promise}
-                    <div>...waiting for data_type</div>
-                  {:then layerType}
-                    {#if layerType.data_type == 'categorical'}
-                      <div class='box' style="background-color: rgb( {color[1][0][0]}, {color[1][0][1]}, {color[1][0][2]} )"> </div>
-                      <div style="display: inline-block;">{color[1][1]}</div><br>
-                    {:else}
-                      <div class='box' style="background-color: rgb( {color[0][0]}, {color[0][1]}, {color[0][2]} )"> </div>
-                      <div style="display: inline-block;">{color[1].toFixed(2)} to {color[2].toFixed(2)} {legend.variable.units}</div><br>
-                    {/if}
-                  {/await}
-                </div>
-              {/each}
-            </div>
-
-          {:catch error}
-            <div style="color: red">{error.message}</div>
-          {/await}
-
-          {#await overlayLayer.openairLink_promise}
-            <div>...waiting for OpenAIRE link</div>
-          {:then openairLink}
-            <div>
-              <a href={openairLink} target="_blank">Link to OpenAIRE metadata &#128279;</a>
-            </div>
-          {:catch error}
-            <div style="color: red">{error.message}</div>
-          {/await}
-
-        </div>
       {/each}
-
     </div>
   {/if}
 </div>
